#182 moved some functions
